+++ src/knnr.jl
@@ -7,7 +7,7 @@ k-Nearest-Neighbours weighted regression (KNNR).
 Keyword arguments:
 * `metric` : Type of dissimilarity used to select the neighbors and to compute the weights 
     (see function `getknn`). Possible values are: `:eucl` (Euclidean), `:mah` (Mahalanobis), 
-    `:sam` (spectral angular distance), `:cor` (correlation distance).
+    `:sam` (a spectral angular distance), `:cos` (cosine distance), `:cor` (a correlation distance).
 * `h` : A scalar defining the shape of the weight function computed by function `winvs`. Lower is h, 
     sharper is the function. See function `winvs` for details (keyword arguments `criw` and `squared` of 
     `winvs` can also be specified here).
@@ -17,15 +17,13 @@ Keyword arguments:
     deviation before the distance and weight computations.
 
 The general principle of this function is as follows (many other variants of kNNR pipelines can be built):
-a) For each new observation to predict, the prediction is the 
-    weighted mean of `y` over a selected neighborhood (in `X`) of 
-    size `k`. 
-b) Within the selected neighborhood, the weights  are defined from 
-    the dissimilarities between the new observation and the neighborhood, 
-    and are computed from function 'winvs'.
+a) For each new observation to predict, the prediction is the weighted mean of `y` over a selected neighborhood
+    (in `X`) of size `k`. 
+b) Within the selected neighborhood, the weights  are defined from the dissimilarities between the new observation 
+    and the neighborhood, and are computed from function 'winvs'.
     
-In general, for X-data with high dimensions, using the Mahalanobis distance requires a preliminary 
-dimensionality reduction (see examples).
+In general, for X-data with high dimensions, using the Mahalanobis distance requires a preliminary dimensionality 
+reduction (see examples).
 
 ## Examples
 ```julia